+++ illuin_inject/bindings/binding.py
@@ -1,0 +1,10 @@
+from typing import Generic, Type
+
+import attr
+
+from illuin_inject.typings import InjectedT
+
+
+@attr.s(auto_attribs=True)
+class Binding(Generic[InjectedT]):
+    target_type: Type[InjectedT]