+++ illuin_inject/dependency_graph/binding_node.py
@@ -1,0 +1,7 @@
+import attr
+
+
+@attr.s(auto_attribs=True)
+class BindingNode:
+    """Represents a binding or a group of bindings in the dependency graph."""
+    pass