+++ illuin_inject/typings.py
@@ -1,0 +1,3 @@
+from typing import Any, TypeVar
+
+InjectedT = TypeVar("InjectedT", bound=Any)