+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 # Changelog
 ## Unreleased
+## 0.1.0
+### Features
+- Added `Injector`
+- Added `BindingSpec`
+- Added `ClassBinding` and `InstanceBinding`
+- Added `SingletonScope`, `PerLookupScope` and `ThreadScope`